Splitt komponenter ut i flere filer
