+++ src/frontend/Sider/Behandling/RevurderFra/Boutgifter/VedtakshistorikkBoutgifter.tsx
@@ -6,7 +6,7 @@ import { Heading, VStack } from '@navikt/ds-react';
 
 import { useHentFullstendigOversikt } from './useHentFullstendigOversikt';
 import DataViewer from '../../../../komponenter/DataViewer';
-import { VedtaksperioderOversiktBoutgifter } from '../../../Personoversikt/Vedtaksperioderoversikt/VedtaksperioderOversiktBoutgifter';
+import { VedtaksperioderOversiktBoutgifter } from '../../../Personoversikt/Vedtaksperioderoversikt/Boutgifter/VedtaksperioderOversiktBoutgifter';
 
 const StyledVStack = styled(VStack)`
     max-width: 50rem;
